GH-382: [refactor] Organize configuration classes into specific packages for better structure
