+++ Microservices/notification-service/src/main/java/app/sportahub/notificationservice/config/auth/SecurityConfig.java
@@ -1,4 +1,4 @@
-package app.sportahub.notificationservice.config;
+package app.sportahub.notificationservice.config.auth;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
