+++ Microservices/notification-service/src/main/java/app/sportahub/notificationservice/config/kafka/KafkaConfig.java
@@ -1,4 +1,4 @@
-package app.sportahub.notificationservice.config;
+package app.sportahub.notificationservice.config.kafka;
 
 import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.clients.producer.ProducerConfig;
